.
*Uploaded the db_bookingsystem.sql database for phpmyadmin. 08/09/2021*

There are two files involved in this project. The 1st one is the Web-Booking-System for the website, and the 2nd one is for its booking system in another repository called "Calendar-Booking-App". Pce linked, these two files are ready to go.

Instructions: 
Download the Web-Booking-System and extract all the files. Change the folder name to 'acdc'.
Download the Calendar-Booking-System and extract all the files. Change the folder name to 'booking'.
Place the folder 'booking' under the 'acdc' folder. So there should be one folder left.
Copy the entire folder 'acdc' and paste it in the directory C: / xampp / htdocs.
Open the XAMPP control panel. Start Apache and MySQL. 
Type in the address 'localhost/phpmyadmin'
Create a database named 'db_bookingsystem'
Go to 'import' and import the db_bookingsystem.sql to your phpmyadmin panel.
Type in the address bar 'localhost/acdc'. Your acdc booking system with website is good to go.